style(modeling) split circle out from ellipse code
